requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2.16
django-behaviors==0.5.1
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
flake8==5.0.4
flake8-annotations==3.0.0
flake8-broken-line==0.6.0
flake8-commas==2.1.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
pep8-naming==0.13.3
Pillow==9.3.0
platformdirs==3.2.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
requests==2.26.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.1
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
